agents/data_analysis_agent.py
from typing import Dict, Any, Optional, List
import pandas as pd
import numpy as np
from .base_agent import BaseAgent

class DataAnalysisAgent(BaseAgent):
    """Agente para análise de dados."""
    
    def __init__(self, **kwargs):
        super().__init__(**kwargs)
    
    @property
    def name(self) -> str:
        return "data_analysis"
    
    @property
    def description(self) -> str:
        return "Agente para análise e visualização de dados"
    
    def run(self, params: Dict[str, Any]) -> Dict[str, Any]:
        """
        Executa análise de dados com base nos parâmetros fornecidos.
        
        Args:
            params: Parâmetros para a análise, incluindo:
                - data_source: Fonte dos dados (arquivo, URL, etc.)
                - analysis_type: Tipo de análise (estatística, clustering, etc.)
                - options: Opções adicionais para a análise
                
        Returns:
            Dict[str, Any]: Resultados da análise
        """
        data_source = params.get("data_source")
        analysis_type = params.get("analysis_type", "statistical")
        options = params.get("options", {})
        
        # Carrega os dados
        df = self._load_data(data_source)
        
        # Executa a análise solicitada
        if analysis_type == "statistical":
            results = self._statistical_analysis(df, options)
        elif analysis_type == "clustering":
            results = self._clustering_analysis(df, options)
        elif analysis_type == "time_series":
            results = self._time_series_analysis(df, options)
        else:
            results = {"error": f"Tipo de análise não suportado: {analysis_type}"}
        
        return results
    
    def _load_data(self, data_source: str) -> pd.DataFrame:
        """
        Carrega dados da fonte especificada.
        
        Args:
            data_source: Caminho do arquivo ou URL
            
        Returns:
            pd.DataFrame: DataFrame com os dados carregados
        """
        try:
            if data_source.endswith(".csv"):
                return pd.read_csv(data_source)
            elif data_source.endswith(".xlsx") or data_source.endswith(".xls"):
                return pd.read_excel(data_source)
            elif data_source.endswith(".json"):
                return pd.read_json(data_source)
            else:
                # Tenta inferir o formato
                return pd.read_csv(data_source)
        except Exception as e:
            print(f"Erro ao carregar dados: {str(e)}")
            return pd.DataFrame()
    
    def _statistical_analysis(self, df: pd.DataFrame, options: Dict[str, Any]) -> Dict[str, Any]:
        """
        Realiza análise estatística básica.
        
        Args:
            df: DataFrame com os dados
            options: Opções para a análise
            
        Returns:
            Dict[str, Any]: Resultados da análise estatística
        """
        results = {
            "summary": df.describe().to_dict(),
            "missing_values": df.isnull().sum().to_dict(),
            "correlation": df.corr().to_dict() if df.select_dtypes(include=[np.number]).shape[1] > 1 else {}
        }
        
        return results
    
    def _clustering_analysis(self, df: pd.DataFrame, options: Dict[str, Any]) -> Dict[str, Any]:
        """
        Realiza análise de clustering.
        
        Args:
            df: DataFrame com os dados
            options: Opções para a análise
            
        Returns:
            Dict[str, Any]: Resultados da análise de clustering
        """
        # Implementação simplificada
        return {"message": "Análise de clustering não implementada completamente"}
    
    def _time_series_analysis(self, df: pd.DataFrame, options: Dict[str, Any]) -> Dict[str, Any]:
        """
        Realiza análise de séries temporais.
        
        Args:
            df: DataFrame com os dados
            options: Opções para a análise
            
        Returns:
            Dict[str, Any]: Resultados da análise de séries temporais
        """
        # Implementação simplificada
        return {"message": "Análise de séries temporais não implementada completamente"}