requirements.txt
# Kognia One - Requisitos Principais

# Frontend
streamlit>=1.22.0
pandas>=1.5.3
plotly>=5.13.1

# API
fastapi>=0.95.0
uvicorn>=0.21.1
pydantic>=1.10.7
python-jose>=3.3.0
passlib>=1.7.4
python-multipart>=0.0.6

# Database
sqlalchemy>=2.0.9
pymysql>=1.0.3
alembic>=1.10.3

# LLMs
openai>=0.27.4
google-generativeai>=0.1.0
anthropic>=0.2.8

# Utilities
python-dotenv>=1.0.0
requests>=2.28.2
pytest>=7.3.1
pytest-cov>=4.1.0

# Desenvolvimento
black>=23.3.0
isort>=5.12.0
flake8>=6.0.0
mypy>=1.2.0